+++ src/environment.rs
@@ -1,7 +1,7 @@
 use crate::lock_file::UvResolutionContext;
 use crate::progress::await_in_progress;
-use crate::project::combine_feature::HasFeatures;
 use crate::project::grouped_environment::GroupedEnvironmentName;
+use crate::project::has_features::HasFeatures;
 use crate::{
     consts, install, install_pypi,
     lock_file::UpdateLockFileOptions,
