+++ src/lock_file/satisfiability.rs
@@ -1,5 +1,5 @@
 use super::{PypiRecord, PypiRecordsByName, RepoDataRecordsByName};
-use crate::project::combine_feature::HasFeatures;
+use crate::project::has_features::HasFeatures;
 use crate::project::manifest::python::{AsPep508Error, RequirementOrEditable};
 use crate::{project::Environment, pypi_marker_env::determine_marker_environment};
 use distribution_types::ParsedGitUrl;
